requirements.txt
asgiref==3.4.1
Django==2.2.10
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==19.3.0
